Add RequestUri to ApiException
If you're making a bunch of calls in a method, any of which can thrown an
ApiException, it's nice to be able to see which request failed without
inspecting the call stack. Specifically I want to be able to log which
request failed.

While making this change I realized that for public exceptions exposed by
the library, I should probably make the constructors internal, since I
don't think there's a use-case for a 3rd party to use the constructor, but
being public it's part of a contract that can't be changed easily. Oh
well.